Separate view section from advanced and translate choices
