+++ src/components/AdvancedQuestionOptions/AdvancedQuestionOptions.tsx
@@ -13,7 +13,6 @@ import UriField from '../FormField/UriField';
 import UndoIcon from '../../images/icons/arrow-undo-outline.svg';
 import './AdvancedQuestionOptions.css';
 import {
-    ICodeSystem,
     IExtentionType,
     IItemProperty,
     IQuestionnaireItemType,
@@ -37,6 +36,7 @@ import GuidanceParam from './Guidance/GuidanceParam';
 import FhirPathSelect from './FhirPathSelect/FhirPathSelect';
 import CalculatedExpression from './CalculatedExpression/CalculatedExpression';
 import CopyFrom from './CopyFrom/CopyFrom';
+import View from './View/view';
 import { createMarkdownExtension, removeItemExtension, setItemExtension } from '../../helpers/extensionHelper';
 import InputField from '../InputField/inputField';
 import {
@@ -53,15 +53,6 @@ import {
 } from '../../helpers/questionTypeFeatures';
 import RadioBtn from '../RadioBtn/RadioBtn';
 import { elementSaveCapability } from '../../helpers/QuestionHelper';
-import {
-    renderingOptions,
-    removeItemCode,
-    addRenderOptionItemCode,
-    RenderingOptionsEnum,
-    choiceRenderOptions,
-    ChoiceRenderOptionCodes,
-    addChoiceRenderOptionItemCode,
-} from '../../helpers/codeHelper';
 
 type AdvancedQuestionOptionsProps = {
     item: QuestionnaireItem;
@@ -149,37 +140,6 @@ const AdvancedQuestionOptions = (props: AdvancedQuestionOptionsProps): JSX.Eleme
     const hasSummaryContainerExtension = isItemControlSummaryContainer(props.item);
 
     const helpTextItem = getHelpTextItem();
-    const checkedRenderOptions = () => {
-        return props.item.extension?.find((ex) => ex.url === IExtentionType.hidden)?.valueBoolean
-            ? RenderingOptionsEnum.Hidden
-            : props.item.code?.find((code) => code.system === ICodeSystem.renderOptionsCodeSystem)?.code ??
-                  RenderingOptionsEnum.None;
-    };
-    const onChangeRenderOptions = (newValue: string) => {
-        removeItemExtension(props.item, IExtentionType.hidden, dispatch);
-        removeItemCode(props.item, ICodeSystem.renderOptionsCodeSystem, dispatch);
-        switch (newValue) {
-            case RenderingOptionsEnum.Hidden:
-                const extension = {
-                    url: IExtentionType.hidden,
-                    valueBoolean: true,
-                };
-                setItemExtension(props.item, extension, dispatch);
-                break;
-            default:
-                addRenderOptionItemCode(props.item, newValue, dispatch);
-                break;
-        }
-    };
-
-    const checkedChoiceRenderOptions = () => {
-        const choiceRendering = props.item.code?.find((code) => code.system === ICodeSystem.choiceRenderOptions);
-        return choiceRendering === undefined ? ChoiceRenderOptionCodes.Default : choiceRendering.code;
-    };
-    const onChangeChoiceRenderOptions = (newValue: string) => {
-        removeItemCode(props.item, ICodeSystem.choiceRenderOptions, dispatch);
-        addChoiceRenderOptionItemCode(props.item, newValue, dispatch);
-    };
 
     const isGroupItemOnGlobalLevel = (groupId: string): boolean => {
         return qOrder.find((i) => i.linkId === groupId) ? true : false;
@@ -464,24 +424,7 @@ const AdvancedQuestionOptions = (props: AdvancedQuestionOptionsProps): JSX.Eleme
                     )}
                 </>
             )}
-            <FormField label={t('View')}>
-                <FormField sublabel={t('Choose if/where the component should be displayed')}>
-                    <RadioBtn
-                        onChange={onChangeRenderOptions}
-                        checked={checkedRenderOptions()}
-                        options={renderingOptions}
-                        name={'elementView-radio'}
-                    />
-                </FormField>
-                <FormField sublabel={t('Choose if all choice answers should be displayed')}>
-                    <RadioBtn
-                        onChange={onChangeChoiceRenderOptions}
-                        checked={checkedChoiceRenderOptions()}
-                        options={choiceRenderOptions}
-                        name={'elementChoiceView-radio'}
-                    />
-                </FormField>
-            </FormField>
+            <View item={props.item} />
             <FormField label={t('Save capabilities')}>
                 <RadioBtn
                     onChange={(newValue: string) => {
